backend/requirements.txt

django==5.0.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
